+++ helpers/image.js
@@ -1,0 +1,34 @@
+function formatImageUrl(url, size) {
+  const isUPPImage = checkUrl(url.binaryUrl);
+  let format;
+
+  if (isUPPImage) {
+    const uuid = extractUUID(url);
+    format = `${process.env.IMAGE_SERVICE_URL}${
+      process.env.REPLACE_IMG_URL
+    }${uuid}`;
+  } else {
+    format = `${process.env.IMAGE_SERVICE_URL}${encodeURIComponent(
+      url.binaryUrl
+    )}`;
+  }
+  return format.concat(`?source=ftlabs&width=${size}`);
+}
+
+function checkUrl(url) {
+  const ftcmsImageRegex = /^https?:\/\/(?:(?:www\.)?ft\.com\/cms|im\.ft-static\.com\/content\/images|com\.ft\.imagepublish\.(?:prod|upp-prod-eu|upp-prod-us)\.s3\.amazonaws\.com|prod-upp-image-read\.ft\.com)\/([a-f\d]{8}-[a-f\d]{4}-[a-f\d]{4}-[a-f\d]{4}-[a-f\d]{12})/g;
+  return ftcmsImageRegex.test(url);
+}
+
+function extractUUID(link) {
+  if (link !== undefined) {
+    const linkWithoutHTTP = link.apiUrl.split("://")[1];
+    return linkWithoutHTTP
+      .replace("api.ft.com/content/", "")
+      .replace("api.ft.com/things/", "");
+  }
+
+  return undefined;
+}
+
+modules.exports = { formatImageUrl };
