+++ routes/spaceUtilisation.js
@@ -19,9 +19,9 @@ router.get("/grid", async (req, res, next) => {
 });
 
 function chunkArray(myArray, chunk_size) {
-  var index = 0;
-  var arrayLength = myArray.length;
-  var tempArray = [];
+  const index = 0;
+  const arrayLength = myArray.length;
+  let tempArray = [];
 
   for (index = 0; index < arrayLength; index += chunk_size) {
     myChunk = myArray.slice(index, index + chunk_size);
