+++ static/css/spaceUtilisation/grid.css
@@ -73,7 +73,7 @@ h1 {
   z-index: 0;
   position: absolute;
   height: 100%;
-  width: 100%;
+  width: auto;
   opacity: 0.7;
 }
