+++ views/index.hbs
@@ -55,20 +55,27 @@
 			</ul>
 		</li>
 		<li>
-			<ul><li>Creating a <a href="https://en.wikipedia.org/wiki/Chord_diagram" target="_blank">Chord diagram</a> to explore the relationships between topic, people, organisation and genre facets</li></ul>
-	    </li>
-	    <li>
+			<ul>
+				<li>Creating a <a href="https://en.wikipedia.org/wiki/Chord_diagram" target="_blank">Chord diagram</a> to explore
+					the relationships between topic, people, organisation and genre facets</li>
+			</ul>
+		</li>
+		<li>
 			<a href="tinder">FT as Tinder</a>
 		<li>
 			<a href="ftMaps">FT Rainbow Maps</a>
-			<ul><li>Using d3 and datamaps to produce a view for articles and topics</li></ul>
+			<ul>
+				<li>Using d3 and datamaps to produce a view for articles and topics</li>
+			</ul>
 		</li>
 		<li>
 			<a href="ftabulous">FTabulous</a>
-			<ul><li>Presenting summarised recent news in a table format</li></ul>
+			<ul>
+				<li>Presenting summarised recent news in a table format</li>
+			</ul>
 		</li>
 		<li>
-			<a href="space">Space utilization</a>
+			<a href="space">Space utilisation</a>
 		</li>
 	</ul>
 </body>
